+++ src/components/About/AboutSession/AboutSection.stories.tsx
@@ -1,12 +1,12 @@
 import { Meta, StoryFn } from '@storybook/react';
-import AboutSession, { AboutSessionProps } from './AboutSession';
+import AboutSection, { AboutSectionProps } from './AboutSection';
 
 import mock from '../../../utils/about.json';
 const about = mock.data.abouts.data[0];
 
 export default {
-  title: 'Components/About/AboutSession',
-  component: AboutSession,
+  title: 'Components/About/AboutSection',
+  component: AboutSection,
   args: {
     about,
   },
@@ -15,6 +15,6 @@ export default {
   },
 } as Meta;
 
-export const Default: StoryFn<AboutSessionProps> = (args) => (
-  <AboutSession {...args} />
+export const Default: StoryFn<AboutSectionProps> = (args) => (
+  <AboutSection {...args} />
 );
