+++ src/components/About/index.ts
@@ -1,2 +1,2 @@
 export * from './AboutImageModal/AboutImageModal';
-export * from './AboutSession/AboutSession';
+export * from './AboutSession/AboutSection';
